feat: implement command execution interface and terminal integration
