+++ app/(main)/instance/console/terminalThemes.ts
@@ -1,0 +1,49 @@
+// OKLCH values taken directly from the design system's color tokens so the terminal
+// stays visually aligned with the app shell without duplicating hex conversions.
+export const DARK_THEME = {
+  background: 'oklch(0.205 0 0)', // --card dark
+  foreground: 'oklch(0.985 0 0)', // --foreground dark
+  cursor: 'oklch(0.922 0 0)',
+  cursorAccent: 'oklch(0.205 0 0)',
+  selectionBackground: 'oklch(0.269 0 0 / 0.5)', // --muted dark
+  black: 'oklch(0.145 0 0)',
+  red: 'oklch(0.704 0.191 22.216)',
+  green: 'oklch(0.696 0.17 162.48)',
+  yellow: 'oklch(0.769 0.188 70.08)',
+  blue: 'oklch(0.488 0.243 264.376)',
+  magenta: 'oklch(0.627 0.265 303.9)',
+  cyan: 'oklch(0.645 0.246 16.439)',
+  white: 'oklch(0.985 0 0)',
+  brightBlack: 'oklch(0.556 0 0)',
+  brightRed: 'oklch(0.704 0.191 22.216)',
+  brightGreen: 'oklch(0.696 0.17 162.48)',
+  brightYellow: 'oklch(0.769 0.188 70.08)',
+  brightBlue: 'oklch(0.488 0.243 264.376)',
+  brightMagenta: 'oklch(0.627 0.265 303.9)',
+  brightCyan: 'oklch(0.645 0.246 16.439)',
+  brightWhite: 'oklch(0.985 0 0)',
+};
+
+export const LIGHT_THEME = {
+  background: 'oklch(1 0 0)', // --card light
+  foreground: 'oklch(0.145 0 0)', // --foreground light
+  cursor: 'oklch(0.205 0 0)',
+  cursorAccent: 'oklch(1 0 0)',
+  selectionBackground: 'oklch(0.97 0 0 / 0.5)', // --muted light
+  black: 'oklch(0.145 0 0)',
+  red: 'oklch(0.577 0.245 27.325)',
+  green: 'oklch(0.6 0.118 184.704)',
+  yellow: 'oklch(0.828 0.189 84.429)',
+  blue: 'oklch(0.398 0.07 227.392)',
+  magenta: 'oklch(0.646 0.222 41.116)',
+  cyan: 'oklch(0.769 0.188 70.08)',
+  white: 'oklch(0.985 0 0)',
+  brightBlack: 'oklch(0.556 0 0)',
+  brightRed: 'oklch(0.577 0.245 27.325)',
+  brightGreen: 'oklch(0.6 0.118 184.704)',
+  brightYellow: 'oklch(0.828 0.189 84.429)',
+  brightBlue: 'oklch(0.398 0.07 227.392)',
+  brightMagenta: 'oklch(0.646 0.222 41.116)',
+  brightCyan: 'oklch(0.769 0.188 70.08)',
+  brightWhite: 'oklch(0.985 0 0)',
+};
